Bug fix: attempts are no longer potentially made to delete non existent settings collection

+++ Solution/WorkingFilesList.ToolWindow/Repository/StoredSettingsRepository.cs
@@ -264,7 +264,12 @@ namespace WorkingFilesList.ToolWindow.Repository
             using (var service = _settingsStoreService.GetSettingsStore(false))
             {
                 var store = (WritableSettingsStore)service.SettingsStore;
-                store.DeleteCollection(_settingsCollectionName);
+                var collectionExists = store.CollectionExists(_settingsCollectionName);
+
+                if (collectionExists)
+                {
+                    store.DeleteCollection(_settingsCollectionName);
+                }
             }
         }
     }
